deleção off-line

+++ app/src/main/java/com/robertolopes/agenda/dao/AlunoDAO.java
@@ -15,7 +15,7 @@ import java.util.UUID;
 
 public class AlunoDAO extends SQLiteOpenHelper {
     public AlunoDAO(Context context) {
-        super(context, "Agenda", null, 5);
+        super(context, "Agenda", null, 6);
     }
 
     @Override
@@ -28,7 +28,8 @@ public class AlunoDAO extends SQLiteOpenHelper {
                 "site TEXT, " +
                 "nota REAL, " +
                 "caminhoFoto TEXT, " +
-                "sincronizado INT DEFAULT 0);";
+                "sincronizado INT DEFAULT 0, " +
+                "desativado INT DEFAULT 0);";
         db.execSQL(sql);
     }
 
@@ -74,6 +75,9 @@ public class AlunoDAO extends SQLiteOpenHelper {
             case 4:
                 String adicionaCampoSicronizado = "ALTER TABLE Alunos ADD COLUMN sincronizado INT DEFAULT 0";
                 db.execSQL(adicionaCampoSicronizado);
+            case 5:
+                String adicionaCampDesativado = "ALTER TABLE Alunos ADD COLUMN desativado INT DEFAULT 0";
+                db.execSQL(adicionaCampDesativado);
         }
     }
 
@@ -101,11 +105,12 @@ public class AlunoDAO extends SQLiteOpenHelper {
         dados.put("nota", aluno.getNota());
         dados.put("caminhoFoto", aluno.getCaminhoFoto());
         dados.put("sincronizado", aluno.getSincronizado());
+        dados.put("desativado", aluno.getDesativado());
         return dados;
     }
 
     public List<Aluno> buscaAlunos() {
-        String sql = "SELECT * FROM Alunos;";
+        String sql = "SELECT * FROM Alunos WHERE desativado = 0;";
         SQLiteDatabase db = getReadableDatabase();
         Cursor c = db.rawQuery(sql, null);
         List<Aluno> alunos = populaAlunos(c);
@@ -125,7 +130,7 @@ public class AlunoDAO extends SQLiteOpenHelper {
             aluno.setNota(c.getDouble(c.getColumnIndex("nota")));
             aluno.setCaminhoFoto(c.getString(c.getColumnIndex("caminhoFoto")));
             aluno.setSincronizado(c.getInt(c.getColumnIndex("sincronizado")));
-
+            aluno.setDesativado(c.getInt(c.getColumnIndex("desativado")));
             alunos.add(aluno);
         }
         c.close();
@@ -143,7 +148,13 @@ public class AlunoDAO extends SQLiteOpenHelper {
         SQLiteDatabase db = getWritableDatabase();
 
         String[] params = {aluno.getId()};
-        db.delete("Alunos", "id = ?", params);
+        if (aluno.estaDesativado()) {
+            db.delete("Alunos", "id = ?", params);
+        } else {
+            aluno.desativa();
+            altera(aluno);
+        }
+
     }
 
     public void altera(Aluno aluno) {
